+++ fedimint-server/src/net/api.rs
@@ -1,62 +1,21 @@
-//! Implements the client API through which users interact with the federation
-use std::cmp::Ordering;
-use std::collections::{BTreeMap, HashMap};
 use std::fmt::{Debug, Formatter};
-use std::path::{Path, PathBuf};
+use std::net::SocketAddr;
+use std::panic::AssertUnwindSafe;
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
-use anyhow::{anyhow, Result};
+use anyhow::Context;
 use async_trait::async_trait;
-use bitcoin_hashes::sha256;
-use fedimint_aead::{encrypt, get_encryption_key, random_salt};
-use fedimint_api_client::api::{
-    FederationStatus, GuardianConfigBackup, PeerConnectionStatus, PeerStatus, StatusResponse,
-};
-use fedimint_core::admin_client::ServerStatus;
-use fedimint_core::backup::{ClientBackupKey, ClientBackupSnapshot};
-use fedimint_core::config::{ClientConfig, JsonWithKind};
-use fedimint_core::core::backup::{SignedBackupRequest, BACKUP_REQUEST_MAX_PAYLOAD_SIZE_BYTES};
-use fedimint_core::core::{DynOutputOutcome, ModuleInstanceId};
-use fedimint_core::db::{
-    Committable, Database, DatabaseTransaction, IDatabaseTransactionOpsCoreTyped,
-};
-use fedimint_core::endpoint_constants::{
-    AUDIT_ENDPOINT, AUTH_ENDPOINT, AWAIT_OUTPUT_OUTCOME_ENDPOINT, AWAIT_SESSION_OUTCOME_ENDPOINT,
-    AWAIT_SIGNED_SESSION_OUTCOME_ENDPOINT, AWAIT_TRANSACTION_ENDPOINT, BACKUP_ENDPOINT,
-    CLIENT_CONFIG_ENDPOINT, FEDERATION_ID_ENDPOINT, GUARDIAN_CONFIG_BACKUP_ENDPOINT,
-    INVITE_CODE_ENDPOINT, MODULES_CONFIG_JSON_ENDPOINT, RECOVER_ENDPOINT,
-    SERVER_CONFIG_CONSENSUS_HASH_ENDPOINT, SESSION_COUNT_ENDPOINT, SESSION_STATUS_ENDPOINT,
-    SHUTDOWN_ENDPOINT, STATUS_ENDPOINT, SUBMIT_TRANSACTION_ENDPOINT, VERSION_ENDPOINT,
-};
-use fedimint_core::epoch::ConsensusItem;
-use fedimint_core::module::audit::{Audit, AuditSummary};
-use fedimint_core::module::registry::ServerModuleRegistry;
-use fedimint_core::module::{
-    api_endpoint, ApiEndpoint, ApiEndpointContext, ApiError, ApiRequestErased, ApiVersion,
-    SerdeModuleEncoding, SupportedApiVersionsSummary,
-};
-use fedimint_core::secp256k1::{PublicKey, SECP256K1};
-use fedimint_core::server::DynServerModule;
-use fedimint_core::session_outcome::{SessionOutcome, SessionStatus, SignedSessionOutcome};
-use fedimint_core::transaction::{SerdeTransaction, Transaction, TransactionError};
-use fedimint_core::{OutPoint, PeerId, TransactionId};
+use fedimint_core::core::ModuleInstanceId;
+use fedimint_core::module::{ApiEndpoint, ApiEndpointContext, ApiError, ApiRequestErased};
 use fedimint_logging::LOG_NET_API;
-use futures::StreamExt;
+use futures::FutureExt;
+use jsonrpsee::server::{PingConfig, RpcServiceBuilder, ServerBuilder, ServerHandle};
+use jsonrpsee::types::ErrorObject;
 use jsonrpsee::RpcModule;
-use tokio::sync::{watch, RwLock};
-use tracing::{debug, info};
+use tracing::{error, info};
 
-use crate::config::io::{
-    CONSENSUS_CONFIG, ENCRYPTED_EXT, JSON_EXT, LOCAL_CONFIG, PRIVATE_CONFIG, SALT_FILE,
-};
-use crate::config::ServerConfig;
-use crate::consensus::engine::get_finished_session_count_static;
-use crate::consensus::transaction::process_transaction_with_dbtx;
-use crate::db::{AcceptedItemPrefix, AcceptedTransactionKey, SignedSessionOutcomeKey};
-use crate::fedimint_core::encoding::Encodable;
-use crate::metrics::{BACKUP_WRITE_SIZE_BYTES, STORED_BACKUPS_COUNT};
-use crate::{check_auth, ApiResult, HasApiContext};
+use crate::metrics;
 
 /// A state that has context for the API, passed to each rpc handler callback
 #[derive(Clone)]
@@ -78,594 +37,110 @@ impl<M: Debug> Debug for RpcHandlerCtx<M> {
     }
 }
 
-#[derive(Clone)]
-pub struct ConsensusApi {
-    /// Our server configuration
-    pub cfg: ServerConfig,
-    /// Database for serving the API
-    pub db: Database,
-    /// Modules registered with the federation
-    pub modules: ServerModuleRegistry,
-    /// Cached client config
-    pub client_cfg: ClientConfig,
-    /// For sending API events to consensus such as transactions
-    pub submission_sender: async_channel::Sender<ConsensusItem>,
-    pub shutdown_sender: watch::Sender<Option<u64>>,
-    pub connection_status_channels: Arc<RwLock<BTreeMap<PeerId, PeerConnectionStatus>>>,
-    pub last_ci_by_peer: Arc<RwLock<BTreeMap<PeerId, u64>>>,
-    pub consensus_status_cache: ExpiringCache<ApiResult<FederationStatus>>,
-    pub supported_api_versions: SupportedApiVersionsSummary,
+/// How long to wait before timing out client connections
+const API_ENDPOINT_TIMEOUT: Duration = Duration::from_secs(60);
+
+/// Has the context necessary for serving API endpoints
+///
+/// Returns the specific `State` the endpoint requires and the
+/// `ApiEndpointContext` which all endpoints can access.
+#[async_trait]
+pub trait HasApiContext<State> {
+    async fn context(
+        &self,
+        request: &ApiRequestErased,
+        id: Option<ModuleInstanceId>,
+    ) -> (&State, ApiEndpointContext<'_>);
 }
 
-impl ConsensusApi {
-    pub fn api_versions_summary(&self) -> &SupportedApiVersionsSummary {
-        &self.supported_api_versions
-    }
+pub type ApiResult<T> = Result<T, ApiError>;
 
-    // we want to return an error if and only if the submitted transaction is
-    // invalid and will be rejected if we were to submit it to consensus
-    pub async fn submit_transaction(
-        &self,
-        transaction: Transaction,
-    ) -> Result<TransactionId, TransactionError> {
-        let txid = transaction.tx_hash();
-
-        debug!(target: LOG_NET_API, %txid, "Received a submitted transaction");
-
-        // Create read-only DB tx so that the read state is consistent
-        let mut dbtx = self.db.begin_transaction_nc().await;
-        // we already processed the transaction before
-        if dbtx
-            .get_value(&AcceptedTransactionKey(txid))
-            .await
-            .is_some()
-        {
-            debug!(target: LOG_NET_API, %txid, "Transaction already accepted");
-            return Ok(txid);
-        }
-
-        // We ignore any writes, as we only verify if the transaction is valid here
-        dbtx.ignore_uncommitted();
-
-        process_transaction_with_dbtx(self.modules.clone(), &mut dbtx, transaction.clone()).await?;
-
-        self.submission_sender
-            .send(ConsensusItem::Transaction(transaction))
-            .await
-            .ok();
-
-        Ok(txid)
-    }
-
-    pub async fn await_transaction(
-        &self,
-        txid: TransactionId,
-    ) -> (Vec<ModuleInstanceId>, DatabaseTransaction<'_, Committable>) {
-        self.db
-            .wait_key_check(&AcceptedTransactionKey(txid), std::convert::identity)
-            .await
-    }
-
-    pub async fn await_output_outcome(
-        &self,
-        outpoint: OutPoint,
-    ) -> Result<SerdeModuleEncoding<DynOutputOutcome>> {
-        let (module_ids, mut dbtx) = self.await_transaction(outpoint.txid).await;
-
-        let module_id = module_ids
-            .into_iter()
-            .nth(outpoint.out_idx as usize)
-            .ok_or(anyhow!("Outpoint index out of bounds {:?}", outpoint))?;
-
-        let outcome = self
-            .modules
-            .get_expect(module_id)
-            .output_status(
-                &mut dbtx.to_ref_with_prefix_module_id(module_id).into_nc(),
-                outpoint,
-                module_id,
-            )
-            .await
-            .expect("The transaction is accepted");
-
-        Ok((&outcome).into())
-    }
-
-    pub async fn session_count(&self) -> u64 {
-        get_finished_session_count_static(&mut self.db.begin_transaction_nc().await).await
-    }
-
-    pub async fn await_signed_session_outcome(&self, index: u64) -> SignedSessionOutcome {
-        self.db
-            .wait_key_check(&SignedSessionOutcomeKey(index), std::convert::identity)
-            .await
-            .0
-    }
-
-    pub async fn session_status(&self, session_index: u64) -> SessionStatus {
-        let mut dbtx = self.db.begin_transaction_nc().await;
-
-        match session_index.cmp(&get_finished_session_count_static(&mut dbtx).await) {
-            Ordering::Greater => SessionStatus::Initial,
-            Ordering::Equal => SessionStatus::Pending(
-                dbtx.find_by_prefix(&AcceptedItemPrefix)
-                    .await
-                    .map(|entry| entry.1)
-                    .collect()
-                    .await,
-            ),
-            Ordering::Less => SessionStatus::Complete(
-                dbtx.get_value(&SignedSessionOutcomeKey(session_index))
-                    .await
-                    .expect("There are no gaps in session outcomes")
-                    .session_outcome,
-            ),
-        }
-    }
-
-    pub async fn get_federation_status(&self) -> ApiResult<FederationStatus> {
-        let peers_connection_status = self.connection_status_channels.read().await.clone();
-        let last_ci_by_peer = self.last_ci_by_peer.read().await.clone();
-        let session_count = self.session_count().await;
-
-        let status_by_peer = peers_connection_status
-            .into_iter()
-            .map(|(peer, connection_status)| {
-                let last_contribution = last_ci_by_peer.get(&peer).cloned();
-                let flagged = last_contribution.unwrap_or(0) + 1 < session_count;
-
-                let consensus_status = PeerStatus {
-                    connection_status,
-                    last_contribution,
-                    flagged,
-                };
-
-                (peer, consensus_status)
-            })
-            .collect::<HashMap<PeerId, PeerStatus>>();
-
-        let peers_flagged = status_by_peer
-            .values()
-            .filter(|status| status.flagged)
-            .count() as u64;
-
-        let peers_online = status_by_peer
-            .values()
-            .filter(|status| status.connection_status == PeerConnectionStatus::Connected)
-            .count() as u64;
-
-        let peers_offline = status_by_peer
-            .values()
-            .filter(|status| status.connection_status == PeerConnectionStatus::Disconnected)
-            .count() as u64;
-
-        Ok(FederationStatus {
-            session_count,
-            peers_online,
-            peers_offline,
-            peers_flagged,
-            status_by_peer,
-        })
-    }
-
-    fn shutdown(&self, index: Option<u64>) {
-        self.shutdown_sender.send_replace(index);
-    }
-
-    async fn get_federation_audit(&self) -> ApiResult<AuditSummary> {
-        let mut dbtx = self.db.begin_transaction_nc().await;
-        // Writes are related to compacting audit keys, which we can safely ignore
-        // within an API request since the compaction will happen when constructing an
-        // audit in the consensus server
-        dbtx.ignore_uncommitted();
-
-        let mut audit = Audit::default();
-        let mut module_instance_id_to_kind: HashMap<ModuleInstanceId, String> = HashMap::new();
-        for (module_instance_id, kind, module) in self.modules.iter_modules() {
-            module_instance_id_to_kind.insert(module_instance_id, kind.as_str().to_string());
-            module
-                .audit(
-                    &mut dbtx.to_ref_with_prefix_module_id(module_instance_id),
-                    &mut audit,
-                    module_instance_id,
-                )
-                .await
-        }
-        Ok(AuditSummary::from_audit(
-            &audit,
-            &module_instance_id_to_kind,
-        ))
-    }
-
-    /// Uses the in-memory config to write a config backup tar archive that
-    /// guardians can download. Private keys are encrypted with the guardian
-    /// password, so it should be safe to store anywhere, this also means the
-    /// backup is useless without the password.
-    async fn get_guardian_config_backup(
-        &self,
-        password: String,
-    ) -> ApiResult<GuardianConfigBackup> {
-        let mut tar_archive_builder = tar::Builder::new(Vec::new());
-
-        let mut append = |name: &Path, data: &[u8]| {
-            let mut header = tar::Header::new_gnu();
-            header.set_path(name).expect("Error setting path");
-            header.set_size(data.len() as u64);
-            header.set_mode(0o644);
-            header.set_cksum();
-            tar_archive_builder
-                .append(&header, data)
-                .expect("Error adding data to tar archive");
-        };
-
-        append(
-            &PathBuf::from(LOCAL_CONFIG).with_extension(JSON_EXT),
-            &serde_json::to_vec(&self.cfg.local).expect("Error encoding local config"),
-        );
-
-        append(
-            &PathBuf::from(CONSENSUS_CONFIG).with_extension(JSON_EXT),
-            &serde_json::to_vec(&self.cfg.consensus).expect("Error encoding consensus config"),
-        );
-
-        // Note that the encrypted config returned here uses a different salt than the
-        // on-disk version. While this may be confusing it shouldn't be a problem since
-        // the content and encryption key are the same. It's unpractical to read the
-        // on-disk version here since the server/api aren't aware of the config dir and
-        // ideally we can keep it that way.
-        let encryption_salt = random_salt();
-        append(&PathBuf::from(SALT_FILE), encryption_salt.as_bytes());
-
-        let private_config_bytes =
-            serde_json::to_vec(&self.cfg.private).expect("Error encoding private config");
-        let encryption_key = get_encryption_key(&password, &encryption_salt)
-            .expect("Generating key from password failed");
-        let private_config_encrypted =
-            hex::encode(encrypt(private_config_bytes, &encryption_key).expect("Encryption failed"));
-        append(
-            &PathBuf::from(PRIVATE_CONFIG).with_extension(ENCRYPTED_EXT),
-            private_config_encrypted.as_bytes(),
-        );
-
-        let tar_archive_bytes = tar_archive_builder
-            .into_inner()
-            .expect("Error building tar archive");
-
-        Ok(GuardianConfigBackup { tar_archive_bytes })
-    }
-
-    async fn handle_backup_request<'s, 'dbtx, 'a>(
-        &'s self,
-        dbtx: &'dbtx mut DatabaseTransaction<'a>,
-        request: SignedBackupRequest,
-    ) -> Result<(), ApiError> {
-        let request = request
-            .verify_valid(SECP256K1)
-            .map_err(|_| ApiError::bad_request("invalid request".into()))?;
-
-        if request.payload.len() > BACKUP_REQUEST_MAX_PAYLOAD_SIZE_BYTES {
-            return Err(ApiError::bad_request("snapshot too large".into()));
-        }
-        debug!(target: LOG_NET_API, id = %request.id, len = request.payload.len(), "Received client backup request");
-        if let Some(prev) = dbtx.get_value(&ClientBackupKey(request.id)).await {
-            if request.timestamp <= prev.timestamp {
-                debug!(id = %request.id, len = request.payload.len(), "Received client backup request with old timestamp - ignoring");
-                return Err(ApiError::bad_request("timestamp too small".into()));
-            }
-        }
-
-        info!(target: LOG_NET_API, id = %request.id, len = request.payload.len(), "Storing new client backup");
-        let overwritten = dbtx
-            .insert_entry(
-                &ClientBackupKey(request.id),
-                &ClientBackupSnapshot {
-                    timestamp: request.timestamp,
-                    data: request.payload.to_vec(),
-                },
-            )
-            .await
-            .is_some();
-        BACKUP_WRITE_SIZE_BYTES.observe(request.payload.len() as f64);
-        if !overwritten {
-            dbtx.on_commit(|| STORED_BACKUPS_COUNT.inc());
-        }
-
+pub fn check_auth(context: &mut ApiEndpointContext) -> ApiResult<()> {
+    if !context.has_auth() {
+        Err(ApiError::unauthorized())
+    } else {
         Ok(())
     }
-
-    async fn handle_recover_request(
-        &self,
-        dbtx: &mut DatabaseTransaction<'_>,
-        id: PublicKey,
-    ) -> Option<ClientBackupSnapshot> {
-        dbtx.get_value(&ClientBackupKey(id)).await
-    }
 }
 
-#[async_trait]
-impl HasApiContext<ConsensusApi> for ConsensusApi {
-    async fn context(
-        &self,
-        request: &ApiRequestErased,
-        id: Option<ModuleInstanceId>,
-    ) -> (&ConsensusApi, ApiEndpointContext<'_>) {
-        let mut db = self.db.clone();
-        let mut dbtx = self.db.begin_transaction().await;
-        if let Some(id) = id {
-            db = self.db.with_prefix_module_id(id);
-            dbtx = dbtx.with_prefix_module_id(id)
+pub async fn spawn<T>(
+    name: &'static str,
+    api_bind: &SocketAddr,
+    module: RpcModule<RpcHandlerCtx<T>>,
+    max_connections: u32,
+) -> ServerHandle {
+    info!(target: LOG_NET_API, "Starting api on ws://{api_bind}");
+
+    ServerBuilder::new()
+        .max_connections(max_connections)
+        .enable_ws_ping(PingConfig::new().ping_interval(Duration::from_secs(10)))
+        .set_rpc_middleware(RpcServiceBuilder::new().layer(metrics::jsonrpsee::MetricsLayer))
+        .build(&api_bind.to_string())
+        .await
+        .context(format!("Bind address: {api_bind}"))
+        .context(format!("API name: {name}"))
+        .expect("Could not build API server")
+        .start(module)
+}
+
+pub fn attach_endpoints<State, T>(
+    rpc_module: &mut RpcModule<RpcHandlerCtx<T>>,
+    endpoints: Vec<ApiEndpoint<State>>,
+    module_instance_id: Option<ModuleInstanceId>,
+) where
+    T: HasApiContext<State> + Sync + Send + 'static,
+    State: Sync + Send + 'static,
+{
+    for endpoint in endpoints {
+        let path = if let Some(module_instance_id) = module_instance_id {
+            // This memory leak is fine because it only happens on server startup
+            // and path has to live till the end of program anyways.
+            Box::leak(format!("module_{}_{}", module_instance_id, endpoint.path).into_boxed_str())
+        } else {
+            endpoint.path
+        };
+        // Check if paths contain any abnormal characters
+        if path.contains(|c: char| !matches!(c, '0'..='9' | 'a'..='z' | '_')) {
+            panic!("Constructing bad path name {path}");
         }
-        (
-            self,
-            ApiEndpointContext::new(
-                db,
-                dbtx,
-                request.auth == Some(self.cfg.private.api_auth.clone()),
-                request.auth.clone(),
-            ),
-        )
-    }
-}
 
-#[async_trait]
-impl HasApiContext<DynServerModule> for ConsensusApi {
-    async fn context(
-        &self,
-        request: &ApiRequestErased,
-        id: Option<ModuleInstanceId>,
-    ) -> (&DynServerModule, ApiEndpointContext<'_>) {
-        let (_, context): (&ConsensusApi, _) = self.context(request, id).await;
-        (
-            self.modules.get_expect(id.expect("required module id")),
-            context,
-        )
-    }
-}
+        // Another memory leak that is fine because the function is only called once at
+        // startup
+        let handler: &'static _ = Box::leak(endpoint.handler);
 
-pub fn server_endpoints() -> Vec<ApiEndpoint<ConsensusApi>> {
-    vec![
-        api_endpoint! {
-            VERSION_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, _v: ()| -> SupportedApiVersionsSummary {
-                Ok(fedimint.api_versions_summary().to_owned())
-            }
-        },
-        api_endpoint! {
-            SUBMIT_TRANSACTION_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, transaction: SerdeTransaction| -> SerdeModuleEncoding<Result<TransactionId, TransactionError>> {
-                let transaction = transaction
-                    .try_into_inner(&fedimint.modules.decoder_registry())
-                    .map_err(|e| ApiError::bad_request(e.to_string()))?;
+        rpc_module
+            .register_async_method(path, move |params, rpc_state| async move {
+                let params = params.one::<serde_json::Value>()?;
+                let rpc_context = &rpc_state.rpc_context;
 
-                // we return an inner error if and only if the submitted transaction is
-                // invalid and will be rejected if we were to submit it to consensus
-                Ok((&fedimint.submit_transaction(transaction).await).into())
-            }
-        },
-        api_endpoint! {
-            AWAIT_TRANSACTION_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, tx_hash: TransactionId| -> TransactionId {
-                debug!(transaction = %tx_hash, "Received request");
+                // Using AssertUnwindSafe here is far from ideal. In theory this means we could
+                // end up with an inconsistent state in theory. In practice most API functions
+                // are only reading and the few that do write anything are atomic. Lastly, this
+                // is only the last line of defense
+                AssertUnwindSafe(tokio::time::timeout(API_ENDPOINT_TIMEOUT, async {
+                    let request = serde_json::from_value(params)
+                        .map_err(|e| ApiError::bad_request(e.to_string()))?;
+                    let (state, context) = rpc_context.context(&request, module_instance_id).await;
 
-                fedimint.await_transaction(tx_hash).await;
-
-                debug!(transaction = %tx_hash, "Sending outcome");
-
-                Ok(tx_hash)
-            }
-        },
-        api_endpoint! {
-            AWAIT_OUTPUT_OUTCOME_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, outpoint: OutPoint| -> SerdeModuleEncoding<DynOutputOutcome> {
-                let outcome = fedimint
-                    .await_output_outcome(outpoint)
-                    .await
-                    .map_err(|e| ApiError::bad_request(e.to_string()))?;
-
-                Ok(outcome)
-            }
-        },
-        api_endpoint! {
-            INVITE_CODE_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context,  _v: ()| -> String {
-                Ok(fedimint.cfg.get_invite_code().to_string())
-            }
-        },
-        api_endpoint! {
-            FEDERATION_ID_ENDPOINT,
-            ApiVersion::new(0, 2),
-            async |fedimint: &ConsensusApi, _context,  _v: ()| -> String {
-                Ok(fedimint.cfg.get_federation_id().to_string())
-            }
-        },
-        api_endpoint! {
-            CLIENT_CONFIG_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, _v: ()| -> ClientConfig {
-                Ok(fedimint.client_cfg.clone())
-            }
-        },
-        api_endpoint! {
-            SERVER_CONFIG_CONSENSUS_HASH_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, _v: ()| -> sha256::Hash {
-                Ok(fedimint.cfg.consensus.consensus_hash())
-            }
-        },
-        api_endpoint! {
-            STATUS_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, _v: ()| -> StatusResponse {
-                let consensus_status = fedimint
-                    .consensus_status_cache
-                    .get(|| fedimint.get_federation_status())
-                    .await?;
-                Ok(StatusResponse {
-                    server: ServerStatus::ConsensusRunning,
-                    federation: Some(consensus_status)
-                })
-            }
-        },
-        api_endpoint! {
-            SESSION_COUNT_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, _v: ()| -> u64 {
-                Ok(fedimint.session_count().await)
-            }
-        },
-        api_endpoint! {
-            AWAIT_SESSION_OUTCOME_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, index: u64| -> SerdeModuleEncoding<SessionOutcome> {
-                Ok((&fedimint.await_signed_session_outcome(index).await.session_outcome).into())
-            }
-        },
-        api_endpoint! {
-            AWAIT_SIGNED_SESSION_OUTCOME_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, index: u64| -> SerdeModuleEncoding<SignedSessionOutcome> {
-                Ok((&fedimint.await_signed_session_outcome(index).await).into())
-            }
-        },
-        api_endpoint! {
-            SESSION_STATUS_ENDPOINT,
-            ApiVersion::new(0, 1),
-            async |fedimint: &ConsensusApi, _context, index: u64| -> SerdeModuleEncoding<SessionStatus> {
-                Ok((&fedimint.session_status(index).await).into())
-            }
-        },
-        api_endpoint! {
-            SHUTDOWN_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, context, index: Option<u64>| -> () {
-                check_auth(context)?;
-                fedimint.shutdown(index);
-                Ok(())
-            }
-        },
-        api_endpoint! {
-            AUDIT_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, context, _v: ()| -> AuditSummary {
-                check_auth(context)?;
-                Ok(fedimint.get_federation_audit().await?)
-            }
-        },
-        api_endpoint! {
-            GUARDIAN_CONFIG_BACKUP_ENDPOINT,
-            ApiVersion::new(0, 2),
-            async |fedimint: &ConsensusApi, context, _v: ()| -> GuardianConfigBackup {
-                check_auth(context)?;
-                let password = context.request_auth().expect("Auth was checked before").0;
-                Ok(fedimint.get_guardian_config_backup(password).await?)
-            }
-        },
-        api_endpoint! {
-            BACKUP_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, context, request: SignedBackupRequest| -> () {
-                fedimint
-                    .handle_backup_request(&mut context.dbtx().into_nc(), request).await?;
-                Ok(())
-
-            }
-        },
-        api_endpoint! {
-            RECOVER_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, context, id: PublicKey| -> Option<ClientBackupSnapshot> {
-                Ok(fedimint
-                    .handle_recover_request(&mut context.dbtx().into_nc(), id).await)
-            }
-        },
-        api_endpoint! {
-            AUTH_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |_fedimint: &ConsensusApi, context, _v: ()| -> () {
-                check_auth(context)?;
-                Ok(())
-            }
-        },
-        api_endpoint! {
-            MODULES_CONFIG_JSON_ENDPOINT,
-            ApiVersion::new(0, 0),
-            async |fedimint: &ConsensusApi, _context, _v: ()| -> BTreeMap<ModuleInstanceId, JsonWithKind> {
-                Ok(fedimint.cfg.consensus.modules_json.clone())
-            }
-        },
-    ]
-}
-
-/// Very simple cache mostly used to protect endpoints against denial of service
-/// attacks
-#[derive(Clone)]
-pub struct ExpiringCache<T> {
-    data: Arc<tokio::sync::Mutex<Option<(T, Instant)>>>,
-    duration: Duration,
-}
-
-impl<T: Clone> ExpiringCache<T> {
-    pub fn new(duration: Duration) -> Self {
-        Self {
-            data: Arc::new(tokio::sync::Mutex::new(None)),
-            duration,
-        }
-    }
-
-    pub async fn get<Fut>(&self, f: impl FnOnce() -> Fut) -> T
-    where
-        Fut: futures::Future<Output = T>,
-    {
-        let mut data = self.data.lock().await;
-        if let Some((data, time)) = data.as_ref() {
-            if time.elapsed() < self.duration {
-                return data.clone();
-            }
-        }
-        let new_data = f().await;
-        *data = Some((new_data.clone(), Instant::now()));
-        new_data
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::time::Duration;
-
-    use fedimint_core::task;
-
-    use crate::net::api::ExpiringCache;
-
-    #[tokio::test]
-    async fn test_expiring_cache() {
-        let cache = ExpiringCache::new(Duration::from_secs(1));
-        let mut counter = 0;
-        let result = cache
-            .get(|| async {
-                counter += 1;
-                counter
+                    (handler)(state, context, request).await
+                }))
+                .catch_unwind()
+                .await
+                .map_err(|_| {
+                    error!(
+                        target: LOG_NET_API,
+                        path, "API handler panicked, DO NOT IGNORE, FIX IT!!!"
+                    );
+                    ErrorObject::owned(500, "API handler panicked", None::<()>)
+                })?
+                .map_err(|tokio::time::error::Elapsed { .. }| {
+                    // TODO: find a better error for this, the error we used before:
+                    // jsonrpsee::core::Error::RequestTimeout
+                    // was moved to be client-side only
+                    ErrorObject::owned(-32000, "Request timeout", None::<()>)
+                })?
+                .map_err(|e| ErrorObject::owned(e.code, e.message, None::<()>))
             })
-            .await;
-        assert_eq!(result, 1);
-        let result = cache
-            .get(|| async {
-                counter += 1;
-                counter
-            })
-            .await;
-        assert_eq!(result, 1);
-        task::sleep(Duration::from_secs(2)).await;
-        let result = cache
-            .get(|| async {
-                counter += 1;
-                counter
-            })
-            .await;
-        assert_eq!(result, 2);
+            .expect("Failed to register async method");
     }
 }
